rework bin structure
